nghttpx: Explicitly cast to uint32_t for hash calculation

+++ src/shrpx_client_handler.cc
@@ -687,7 +687,9 @@ uint32_t compute_affinity_from_ip(const StringRef &ip) {
     return util::hash32(ip);
   }
 
-  return (buf[0] << 24) | (buf[1] << 16) | (buf[2] << 8) | buf[3];
+  return (static_cast<uint32_t>(buf[0]) << 24) |
+         (static_cast<uint32_t>(buf[1]) << 16) |
+         (static_cast<uint32_t>(buf[2]) << 8) | static_cast<uint32_t>(buf[3]);
 }
 } // namespace
